Add persistent ephemeral node support to metadata store

+++ kaldb/src/main/java/com/slack/kaldb/metadata/core/KaldbMetadataStore.java
@@ -3,16 +3,21 @@ package com.slack.kaldb.metadata.core;
 import static com.slack.kaldb.server.KaldbConfig.DEFAULT_ZK_TIMEOUT_SECS;
 
 import com.slack.kaldb.util.RuntimeHalterImpl;
+import io.micrometer.core.instrument.Counter;
+import io.micrometer.core.instrument.MeterRegistry;
 import java.io.Closeable;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionException;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
+import org.apache.curator.framework.recipes.nodes.PersistentNode;
 import org.apache.curator.x.async.AsyncCuratorFramework;
 import org.apache.curator.x.async.api.CreateOption;
 import org.apache.curator.x.async.modeled.ModelSerializer;
@@ -23,17 +28,30 @@ import org.apache.curator.x.async.modeled.cached.CachedModeledFramework;
 import org.apache.curator.x.async.modeled.cached.ModeledCacheListener;
 import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.data.Stat;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * KaldbMetadataStore is a class which provides consistent ZK apis for all the metadata store class.
  *
  * <p>Every method provides an async and a sync API. In general, use the async API you are
  * performing batch operations and a sync if you are performing a synchronous operation on a node.
+ *
+ * <p><a href="https://curator.apache.org/docs/recipes-persistent-node">Persistent node recipie</a>
  */
 public class KaldbMetadataStore<T extends KaldbMetadata> implements Closeable {
+  private static final Logger LOG = LoggerFactory.getLogger(KaldbMetadataStore.class);
+
+  public static String PERSISTENT_EPHEMERAL_PROPERTY = "kaldb.metadata.persistentEphemeral";
   protected final String storeFolder;
 
   private final ZPath zPath;
+
+  private final CreateMode createMode;
+
+  private final AsyncCuratorFramework curator;
+
+  private final ModelSpec<T> modelSpec;
 
   private final CountDownLatch cacheInitialized = new CountDownLatch(1);
 
@@ -44,17 +62,26 @@ public class KaldbMetadataStore<T extends KaldbMetadata> implements Closeable {
   private final Map<KaldbMetadataStoreChangeListener<T>, ModeledCacheListener<T>> listenerMap =
       new ConcurrentHashMap<>();
 
+  private final Map<String, PersistentNode> persistentNodeMap = new ConcurrentHashMap<>();
+
+  public static final String PERSISTENT_NODE_RECREATED_COUNTER =
+      "metadata_persistent_node_recreated";
+  private final Counter persistentNodeRecreatedCounter;
+
   public KaldbMetadataStore(
       AsyncCuratorFramework curator,
       CreateMode createMode,
       boolean shouldCache,
       ModelSerializer<T> modelSerializer,
-      String storeFolder) {
+      String storeFolder,
+      MeterRegistry meterRegistry) {
 
+    this.createMode = createMode;
+    this.curator = curator;
     this.storeFolder = storeFolder;
     this.zPath = ZPath.parseWithIds(String.format("%s/{name}", storeFolder));
 
-    ModelSpec<T> modelSpec =
+    this.modelSpec =
         ModelSpec.builder(modelSerializer)
             .withPath(zPath)
             .withCreateOptions(
@@ -70,11 +97,54 @@ public class KaldbMetadataStore<T extends KaldbMetadata> implements Closeable {
     } else {
       cachedModeledFramework = null;
     }
+
+    persistentNodeRecreatedCounter = meterRegistry.counter(PERSISTENT_NODE_RECREATED_COUNTER);
+    LOG.info(
+        "Persistent ephemeral mode '{}' enabled - {}",
+        PERSISTENT_EPHEMERAL_PROPERTY,
+        persistentEphemeralModeEnabled());
+  }
+
+  public static boolean persistentEphemeralModeEnabled() {
+    return Boolean.parseBoolean(System.getProperty(PERSISTENT_EPHEMERAL_PROPERTY, "true"));
   }
 
   public CompletionStage<String> createAsync(T metadataNode) {
-    // by passing the version 0, this will throw if we attempt to create and it already exists
-    return modeledClient.set(metadataNode, 0);
+    if (createMode == CreateMode.EPHEMERAL && persistentEphemeralModeEnabled()) {
+      String nodePath = resolvePath(metadataNode);
+      // persistent node already implements NodeDataChanged
+      PersistentNode node =
+          new PersistentNode(
+              curator.unwrap(),
+              createMode,
+              false,
+              nodePath,
+              modelSpec.serializer().serialize(metadataNode));
+      persistentNodeMap.put(nodePath, node);
+      node.start();
+      // todo - what happens when attempting to create over existing?
+      return CompletableFuture.supplyAsync(
+          () -> {
+            try {
+              node.waitForInitialCreate(DEFAULT_ZK_TIMEOUT_SECS, TimeUnit.SECONDS);
+              node.getListenable().addListener(_ -> persistentNodeRecreatedCounter.increment());
+              return nodePath;
+            } catch (Exception e) {
+              throw new CompletionException(e);
+            }
+          });
+    } else {
+      // by passing the version 0, this will throw if we attempt to create and it already exists
+      return modeledClient.set(metadataNode, 0);
+    }
+  }
+
+  // resolveForSet
+  private String resolvePath(T model) {
+    if (modelSpec.path().isResolved()) {
+      return modelSpec.path().fullPath();
+    }
+    return modelSpec.path().resolved(model).fullPath();
   }
 
   public void createSync(T metadataNode) {
@@ -120,7 +190,18 @@ public class KaldbMetadataStore<T extends KaldbMetadata> implements Closeable {
   }
 
   public CompletionStage<Stat> updateAsync(T metadataNode) {
-    return modeledClient.update(metadataNode);
+    PersistentNode node = getPersistentNodeIfExists(metadataNode);
+    if (node != null) {
+      try {
+        node.waitForInitialCreate(DEFAULT_ZK_TIMEOUT_SECS, TimeUnit.SECONDS);
+        node.setData(modelSpec.serializer().serialize(metadataNode));
+        return CompletableFuture.completedFuture(null);
+      } catch (Exception e) {
+        throw new CompletionException(e);
+      }
+    } else {
+      return modeledClient.update(metadataNode);
+    }
   }
 
   public void updateSync(T metadataNode) {
@@ -134,7 +215,18 @@ public class KaldbMetadataStore<T extends KaldbMetadata> implements Closeable {
   }
 
   public CompletionStage<Void> deleteAsync(String path) {
-    return modeledClient.withPath(zPath.resolved(path)).delete();
+    PersistentNode node = removePersistentNodeIfExists(path);
+    if (node != null) {
+      try {
+        node.waitForInitialCreate(DEFAULT_ZK_TIMEOUT_SECS, TimeUnit.SECONDS);
+        node.close();
+        return CompletableFuture.completedFuture(null);
+      } catch (Exception e) {
+        throw new CompletionException(e);
+      }
+    } else {
+      return modeledClient.withPath(zPath.resolved(path)).delete();
+    }
   }
 
   public void deleteSync(String path) {
@@ -146,7 +238,18 @@ public class KaldbMetadataStore<T extends KaldbMetadata> implements Closeable {
   }
 
   public CompletionStage<Void> deleteAsync(T metadataNode) {
-    return modeledClient.withPath(zPath.resolved(metadataNode)).delete();
+    PersistentNode node = removePersistentNodeIfExists(metadataNode);
+    if (node != null) {
+      try {
+        node.waitForInitialCreate(DEFAULT_ZK_TIMEOUT_SECS, TimeUnit.SECONDS);
+        node.close();
+        return CompletableFuture.completedFuture(null);
+      } catch (Exception e) {
+        throw new CompletionException(e);
+      }
+    } else {
+      return modeledClient.withPath(zPath.resolved(metadataNode)).delete();
+    }
   }
 
   public void deleteSync(T metadataNode) {
@@ -210,6 +313,18 @@ public class KaldbMetadataStore<T extends KaldbMetadata> implements Closeable {
     }
   }
 
+  private PersistentNode getPersistentNodeIfExists(T metadataNode) {
+    return persistentNodeMap.getOrDefault(resolvePath(metadataNode), null);
+  }
+
+  private PersistentNode removePersistentNodeIfExists(T metadataNode) {
+    return persistentNodeMap.remove(resolvePath(metadataNode));
+  }
+
+  private PersistentNode removePersistentNodeIfExists(String path) {
+    return persistentNodeMap.remove(zPath.resolved(path).fullPath());
+  }
+
   private ModeledCacheListener<T> getCacheInitializedListener() {
     return new ModeledCacheListener<T>() {
       @Override
@@ -227,9 +342,17 @@ public class KaldbMetadataStore<T extends KaldbMetadata> implements Closeable {
 
   @Override
   public void close() {
+    persistentNodeMap.forEach(
+        (_, persistentNode) -> {
+          try {
+            persistentNode.close();
+          } catch (Exception ignored) {
+          }
+        });
+
     if (cachedModeledFramework != null) {
       listenerMap.forEach(
-          (kaldbMetadataStoreChangeListener, tModeledCacheListener) ->
+          (_, tModeledCacheListener) ->
               cachedModeledFramework.listenable().removeListener(tModeledCacheListener));
       cachedModeledFramework.close();
     }
